Non-functional code-style fixes
