+++ src/DurableTask.AzureStorage/Utils.cs
@@ -84,38 +84,4 @@ namespace DurableTask.AzureStorage
             return historyEvents.Count(e => e.EventType == EventType.OrchestratorStarted);
         }
     }
-
-    /// <summary>
-    /// Class to hold statistics about this execution of purge history
-    /// </summary>
-    public class PurgeHistoryResult
-    {
-        /// <summary>
-        /// Constructor for purge history statistics
-        /// </summary>
-        /// <param name="storageRequests">Requests sent to storage</param>
-        /// <param name="instancesDeleted">Number of instances deleted</param>
-        /// <param name="rowsDeleted">Number of rows deleted</param>
-        public PurgeHistoryResult(int storageRequests, int instancesDeleted, int rowsDeleted)
-        {
-            this.StorageRequests = storageRequests;
-            this.InstancesDeleted = instancesDeleted;
-            this.RowsDeleted = rowsDeleted;
-        }
-
-        /// <summary>
-        /// Number of requests sent to Storage during this execution of purge history
-        /// </summary>
-        public int StorageRequests { get; }
-
-        /// <summary>
-        /// Number of instances deleted during this execution of purge history
-        /// </summary>
-        public int InstancesDeleted { get; }
-
-        /// <summary>
-        /// Number of rows deleted during this execution of purge history
-        /// </summary>
-        public int RowsDeleted { get; }
-    }
 }
